# docs/swagger.yaml
definitions:
...
  models.Address:
    properties:
      city:
        type: string
      contact_id:
        type: integer
      country:
        type: string
      extended_street:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      label_type_id:
        type: integer
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
      type_label:
        type: string
    type: object
...
  models.Contact:
    properties:
      addresses:
        items:
          $ref: '#/definitions/models.Address'
        type: array
      anniversary:
        example: "2022-01-03T00:00:00Z"
        type: string
      anniversary_day:
        description: 1-31, for partial dates
        example: 3
        type: integer
      anniversary_month:
        description: 1-12, for partial dates
        example: 1
        type: integer
      avatar_base64:
        type: string
      avatar_mime_type:
        type: string
      birthday:
        example: "1990-12-15T00:00:00Z"
        type: string
      birthday_day:
        description: 1-31, for partial dates
        example: 15
        type: integer
      birthday_month:
        description: 1-12, for partial dates
        example: 12
        type: integer
      created_at:
        type: string
      deletedAt:
        type: string
      emails:
        items:
          $ref: '#/definitions/models.Email'
        type: array
      etag:
        type: string
      exclude_from_sync:
        type: boolean
      family_name:
        example: Doe
        type: string
      full_name:
        description: Computed from name parts
        example: Dr. John Frank Doe III
        type: string
      gender:
        description: M, F, O, N, U
        example: M
        type: string
      given_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      last_modified_token:
        type: integer
      maiden_name:
        example: Parks
        type: string
      metadata:
        type: string
      middle_name:
        example: Frank
        type: string
      nickname:
        example: broheim
        type: string
      notes:
        example: Met at work conference 2023
        type: string
      organizations:
        items:
          $ref: '#/definitions/models.Organization'
        type: array
      other_dates:
        items:
          $ref: '#/definitions/models.OtherDate'
        type: array
      other_relationships:
        items:
          $ref: '#/definitions/models.OtherRelationship'
        type: array
      phones:
        items:
          $ref: '#/definitions/models.Phone'
        type: array
      phonetic_first_name:
        example: Par-cor
        type: string
      phonetic_last_name:
        example: Par-cor
        type: string
      phonetic_middle_name:
        example: Par-cor
        type: string
      prefix:
        example: Dr.
        type: string
      pronunciation_first_name:
        example: Par-cor
        type: string
      pronunciation_last_name:
        example: Par-cor
        type: string
      relationships:
        items:
          $ref: '#/definitions/models.Relationship'
        type: array
      suffix:
        example: III
        type: string
      uid:
        example: "42"
        type: string
      updated_at:
        type: string
      urls:
        items:
          $ref: '#/definitions/models.URL'
        type: array
      user_id:
        type: integer
      version_token:
        type: integer
    type: object
...
  models.Email:
    properties:
      contact_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      label_type_id:
        type: integer
      type_label:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: validation_error
        type: string
      field:
        example: full_name
        type: string
      message:
        example: Full name is required
        type: string
    type: object
...
  models.Organization:
    properties:
      contact_id:
        type: integer
      department:
        type: string
      id:
        type: integer
      is_primary:
        description: not yet used
        type: boolean
      name:
        type: string
      phonetic_name:
        type: string
      title:
        type: string
    type: object
...
  models.OtherDate:
    properties:
      contact_id:
        type: integer
      event_date:
        type: string
      event_date_day:
        description: 1-31, for partial dates
        type: integer
      event_date_month:
        description: 1-12, for partial dates
        type: integer
      event_name:
        type: string
      id:
        type: integer
    type: object
...
  models.OtherRelationship:
    properties:
      contact:
        $ref: '#/definitions/models.Contact'
      contact_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      related_contact_name:
        type: string
      relationship_name:
        type: string
    type: object
  models.Phone:
    properties:
      contact_id:
        type: integer
      id:
        type: integer
      is_primary:
        type: boolean
      label_type_id:
        type: integer
      last_formatted_at:
        type: string
      phone:
        type: string
      type_label:
        type: string
    type: object
...
  models.Relationship:
    properties:
      contact:
        $ref: '#/definitions/models.Contact'
      contact_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      related_contact:
        $ref: '#/definitions/models.Contact'
      related_contact_id:
        type: integer
      relationship_type:
        $ref: '#/definitions/models.RelationshipType'
    type: object
  models.RelationshipType:
    properties:
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      reverse_name_female:
        type: string
      reverse_name_male:
        type: string
      reverse_name_neutral:
        type: string
    type: object
...
  models.URL:
    properties:
      contact_id:
        type: integer
      id:
        type: integer
      is_primary:
        type: boolean
      label_type_id:
        type: integer
      type_label:
        type: string
      url:
        type: string
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/steveredden/KindredCard/issues
  description: Personal CRM for managing contacts, relationships, and important dates
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.txt
  title: KindredCard API
  version: "1.0"
paths:
...
  /api/v1/contacts:
    post:
      consumes:
      - application/json
      description: Create a new contact with the provided information
      parameters:
      - description: Contact information
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.Contact'
      produces:
      - application/json
      responses:
        "201":
          description: Created contact
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Contact already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Create a new contact
      tags:
      - contacts
...
  /contacts:
    get:
      consumes:
      - application/json
      description: Get all contacts for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Contact'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Lists all contacts
      tags:
      - contacts
...
securityDefinitions:
  ApiTokenAuth:
    description: API token for authentication
    in: header
    name: session
    type: apiKey
swagger: "2.0"